enable more tests
